refactor reports to sub-package
